Move out UTs to their own files
